add a few validation
