+++ client/package.json
@@ -24,6 +24,7 @@
         "react-responsive-carousel": "^3.2.23",
         "react-router-dom": "^6.8.1",
         "react-scripts": "5.0.1",
+        "react-select": "^5.7.0",
         "redux-persist": "^6.0.0",
         "web-vitals": "^2.1.4",
         "yup": "^1.0.0"
