+++ client/src/components/CartPage/CartPage.jsx
@@ -3,7 +3,6 @@ import { Link } from 'react-router-dom';
 import BreadCrumbs from '../BreadCrumbs/BreadCrumbs';
 import styles from './CartPage.module.scss';
 import ShoppingBag from 'components/ShoppingBag';
-import Form from './Form';
 import CarouselSection from 'components/CarouselSection';
 import { useSelector } from 'react-redux';
 import getAllProducts from 'api/getAllProducts';
@@ -37,7 +36,6 @@ function CartPage() {
             ) : (
                 <>
                     <ShoppingBag />
-                    <Form/>
                     <Link to="/checkout" className={styles.checkoutBtn}>
                         PROCEED TO CHECKOUT
                     </Link>
