--- client/src/components/CartPage/Form/Form.module.scss
@@ -1,53 +1,0 @@
-.red {
-    color: red;
-}
-.input_wrapper {
-    display: flex;
-
-    justify-content: space-between;
-}
-.checkbox_wrapper {
-    display: flex;
-    padding-top: 3rem;
-}
-.form {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    padding-top: 2rem;
-    padding-bottom: 2rem;
-    margin-top: 2rem;
-    margin-bottom: 3rem;
-    border-top: 1px solid #373F41;
-    border-bottom: 1px solid #373F41;
-}
-.shipping_title {
-    font-family: 'Mulish';
-    font-style: normal;
-    font-weight: 400;
-    color: #373f41;
-    font-size: 2.4rem;
-    line-height: 2rem;
-    text-transform: uppercase;
-}
-.inputs_wrapper {
-    padding-bottom: 2rem;
-    border-bottom: 1px solid #373F41;
-}
-.input:first-child{
-  margin-right: 2rem;
-}
-.input{
-border: 1px solid #8C8C8C ;
-margin-bottom: 2rem;
-width: 100%;
-border-radius:  0.5rem;
-color: #8C8C8C;
-}
-.input_area{
-  width: 100%;
-  padding: 5px;
-}
-.select{
-  border-right: 1px solid black;
-}
